Add RecordSender Mode; Add JobResults Function
Add new RecordSender mode that adds support for an arbitrary function
that can send records to be used for the bulk job.

Add JobResults function to get all results.
